Corrigindo todos os outros controllers e services para trabalhar com o tratador de erro global.

+++ src/main/java/com/yokota/treino/service/WorkoutService.java
@@ -9,6 +9,7 @@ import com.yokota.treino.dtos.workout.WorkoutResponseDTO;
 import com.yokota.treino.model.worksheet.Worksheet;
 import com.yokota.treino.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -47,7 +48,10 @@ public class WorkoutService {
     }
 
     //Delete o treino
-    public void deleteWorkout(Workout workout){
+    public void deleteWorkout(Workout workout, User user){
+        if(!workout.getUser().equals(user)){
+            throw new AccessDeniedException("You must be owner of this workout");
+        }
         workoutRepository.delete(workout);
     }
